sj-trading/pyproject.toml
[project]
name = "sj-trading"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "polars-talib>=0.1.3",
    "polars>=1.9.0",
    "python-dotenv>=1.0.1",
    "shioaji[speed]>=1.2.5",
    "typer>=0.9.0",
]

[project.scripts]
sj-trading = "sj_trading.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "ipykernel>=6.29.5",
]